Disable some mouse inputs and try to get menu music working
No more mouse inputs on selectmusic and titlescreen. The inputs still work on pack downloader and bundles but the cursor is invisible. It will stay like this until I find a way to make it interactable with keyboard only. As for the menu music it's kinda fucked but maybe I'll figure that out sometime